[APINotes] Colocate diagnostics FileCheck lines with YAML
