--- clang/test/APINotes/Inputs/Headers/WhereParametersDiagnostics.apinotes
@@ -1,50 +1,0 @@
----
-Name: WhereParametersDiagnostics
-Functions:
-- Name: unmatchedGlobal
-  Where:
-    Parameters:
-    - int
-  SwiftName: shouldNotApplyGlobal(_:)
-- Name: diagnosticBroadGlobal
-  SwiftPrivate: true
-- Name: diagnosticBroadGlobal
-  Where:
-    Parameters:
-    - int
-  SwiftName: shouldNotApplyBroadGlobal(_:)
-- Name: diagnosticMatchedGlobal
-  Where:
-    Parameters:
-    - int
-  SwiftName: diagnosticMatchedGlobal(_:)
-- Name: diagnosticAliasMatchedGlobal
-  Where:
-    Parameters:
-    - int
-  SwiftName: diagnosticAliasMatchedGlobal(_:)
-Tags:
-- Name: DiagnosticWidget
-  Methods:
-  - Name: unmatchedMethod
-    Where:
-      Parameters:
-      - int
-    SwiftName: shouldNotApplyMethod(_:)
-  - Name: diagnosticBroadMethod
-    SwiftPrivate: true
-  - Name: diagnosticBroadMethod
-    Where:
-      Parameters:
-      - int
-    SwiftName: shouldNotApplyBroadMethod(_:)
-  - Name: diagnosticMatchedMethod
-    Where:
-      Parameters:
-      - int
-    SwiftName: diagnosticMatchedMethod(_:)
-  - Name: diagnosticAliasMatchedMethod
-    Where:
-      Parameters:
-      - int
-    SwiftName: diagnosticAliasMatchedMethod(_:)
