--- clang/test/APINotes/Inputs/Headers/WhereParametersDiagnostics.h
@@ -1,18 +1,0 @@
-#ifndef WHERE_PARAMETERS_DIAGNOSTICS_H
-#define WHERE_PARAMETERS_DIAGNOSTICS_H
-
-using DiagnosticAliasInt = int;
-
-void unmatchedGlobal(float);
-void diagnosticBroadGlobal(float);
-void diagnosticMatchedGlobal(int);
-void diagnosticAliasMatchedGlobal(DiagnosticAliasInt);
-
-struct DiagnosticWidget {
-  void unmatchedMethod(float);
-  void diagnosticBroadMethod(float);
-  void diagnosticMatchedMethod(int);
-  void diagnosticAliasMatchedMethod(DiagnosticAliasInt);
-};
-
-#endif // WHERE_PARAMETERS_DIAGNOSTICS_H
