+++ clang/test/APINotes/Inputs/Headers/module.modulemap
@@ -75,8 +75,3 @@ module WhereParametersSema {
   header "WhereParametersSema.h"
   export *
 }
-
-module WhereParametersDiagnostics {
-  header "WhereParametersDiagnostics.h"
-  export *
-}
